Refactor zip extraction functionality into separate utility module
- Moved ZipExtractThread and ExtractionProgressWindow classes from app/views/main_content_panel.py to a new app/utils/zip_extractor.py file.
- Updated imports in main_content_panel.py to reference the new utility module.
 This refactoring improves code organization by separating zip extraction logic into its own dedicated file.
